# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AI Life OS backend — tasks, notes, conversations, and AI chat.",
        "title": "Jarvis API",
        "contact": {
            "name": "Mahdi"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/v1",
    "paths": {
...
        "/tasks": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "List all tasks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response-array_models_Task"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response-any"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Create a task",
                "parameters": [
                    {
                        "description": "Task to create",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Task"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response-models_Task"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response-any"
                        }
                    }
                }
            }
        },
        "/tasks/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Get a task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response-models_Task"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response-any"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Delete a task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response-any"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Update a task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated task fields",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Task"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response-models_Task"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response-any"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.Pagination": {
            "type": "object",
            "properties": {
                "page": {
                    "description": "Page is the current page number (1‑based).",
                    "type": "integer"
                },
                "page_size": {
                    "description": "PageSize is the maximum number of items per page.",
                    "type": "integer"
                },
                "total_items": {
                    "description": "TotalItems is the total number of items across all pages.",
                    "type": "integer"
                },
                "total_pages": {
                    "description": "TotalPages is the total number of pages (TotalItems / PageSize).",
                    "type": "integer"
                }
            }
        },
        "controllers.Response-any": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/controllers.ResponseError"
                },
                "meta": {
                    "$ref": "#/definitions/controllers.ResponseMeta"
                }
            }
        },
...
        "controllers.Response-array_models_Task": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Task"
                    }
                },
                "error": {
                    "$ref": "#/definitions/controllers.ResponseError"
                },
                "meta": {
                    "$ref": "#/definitions/controllers.ResponseMeta"
                }
            }
        },
...
        "controllers.Response-models_Task": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Task"
                },
                "error": {
                    "$ref": "#/definitions/controllers.ResponseError"
                },
                "meta": {
                    "$ref": "#/definitions/controllers.ResponseMeta"
                }
            }
        },
        "controllers.ResponseError": {
            "type": "object",
            "properties": {
                "fields": {
                    "description": "Fields contains per‑field validation errors (usually for 422 responses).",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controllers.ResponseErrorField"
                    }
                },
                "message": {
                    "description": "Message is a human‑readable description of the error.",
                    "type": "string"
                }
            }
        },
        "controllers.ResponseErrorField": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                }
            }
        },
        "controllers.ResponseMeta": {
            "type": "object",
            "properties": {
                "pagination": {
                    "description": "Pagination contains cursor/offset‑based pagination info.\nIt is nil for endpoints that do not paginate.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/controllers.Pagination"
                        }
                    ]
                },
                "request_id": {
                    "description": "RequestID can be used for tracing requests across services.",
                    "type": "string"
                },
                "warnings": {
                    "description": "Warnings contains non‑fatal messages (e.g. “this endpoint is deprecated”).",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "models.Tag": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Task": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "doneAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "parentID": {
                    "type": "string"
                },
                "scheduledFrom": {
                    "type": "string"
                },
                "scheduledTo": {
                    "type": "string"
                },
                "score": {
                    "type": "number"
                },
                "scoreUnit": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Tag"
                    }
                },
                "targetScore": {
                    "description": "Target score for binary (done/undone) task is 1",
                    "type": "number",
                    "minimum": 0
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    }
}